refactor!: Remove special syntax for a rules with `error` severity
BREAKING CHANGE: As a part of the migration to ESLint 9 and making things more explicit, we decided to drop the special syntax from the configurations.
Some rules that previously had a severity of “warning” now have a severity of “error”.
Although most of them were fixed automatically, something unexpected can happen.
Autofix of such rules is still working.